+++ src/Schema/test.js
@@ -29,6 +29,7 @@ describe('Schema', () => {
             col1: { name: 'col1', type: 'string' },
             col2: { name: 'col2', type: 'number' },
           },
+          columnArray: [{ name: 'col1', type: 'string' }, { name: 'col2', type: 'number' }],
         },
         bar: {
           name: 'bar',
@@ -37,6 +38,11 @@ describe('Schema', () => {
             col2: { name: 'col2', type: 'boolean' },
             col3: { name: 'col3', type: 'boolean' },
           },
+          columnArray: [
+            { name: 'col1', type: 'number' },
+            { name: 'col2', type: 'boolean' },
+            { name: 'col3', type: 'boolean' },
+          ],
         },
       },
     })
@@ -60,5 +66,19 @@ describe('Schema', () => {
         columns: [{ name: 'last_modified', type: 'string' }],
       }),
     ).toThrow(/last_modified must be.*number/)
+  })
+  it('does not allow unsafe characters', () => {
+    expect(() =>
+      tableSchema({
+        name: 'foo',
+        columns: [{ name: '"hey"', type: 'string' }],
+      }),
+    ).toThrow(/must contain only safe characters/)
+    expect(() =>
+      tableSchema({
+        name: 'foo\' and delete * from users --',
+        columns: [{ name: 'hey', type: 'string' }],
+      }),
+    ).toThrow(/must contain only safe characters/)
   })
 })