+++ src/adapters/lokijs/worker/executor.js
@@ -4,7 +4,7 @@ import Loki, { LokiCollection, type LokiMemoryAdapter } from 'lokijs'
 import { prop, forEach, values } from 'rambdax'
 import { logger } from '../../../utils/common'
 
-import type { CachedQueryResult, CachedFindResult } from '../../type'
+import type { CachedQueryResult, CachedFindResult, BatchOperation } from '../../type'
 import type { TableName, AppSchema, SchemaVersion, TableSchema } from '../../../Schema'
 import type {
   SchemaMigrations,
@@ -19,7 +19,7 @@ import { type RawRecord, sanitizedRaw, setRawSanitized, type DirtyRaw } from '..
 
 import { newLoki, deleteDatabase } from './lokiExtensions'
 import executeQuery from './executeQuery'
-import type { WorkerBatchOperation } from '../common'
+
 import type { LokiAdapterOptions } from '../index'
 
 const SCHEMA_VERSION_KEY = '_loki_schema_version'
@@ -98,11 +98,6 @@ export default class LokiExecutor {
     return executeQuery(query, this.loki).count()
   }
 
-  create(table: TableName<any>, raw: RawRecord): void {
-    this.loki.getCollection(table).insert(raw)
-    this.markAsCached(table, raw.id)
-  }
-
   update(table: TableName<any>, rawRecord: RawRecord): void {
     const collection = this.loki.getCollection(table)
     // Loki identifies records using internal $loki ID so we must find the saved record first
@@ -129,23 +124,49 @@ export default class LokiExecutor {
     }
   }
 
-  batch(operations: WorkerBatchOperation[]): void {
+  batch(operations: BatchOperation[]): void {
     // TODO: Only add to cached records if all is successful
     // TODO: Transactionality
+
+    const recordsToCreate: { [TableName<any>]: RawRecord[] } = {}
+
     operations.forEach(operation => {
       const [type, table, raw] = operation
       switch (type) {
         case 'create':
-          this.create(table, raw)
+          if (!recordsToCreate[table]) {
+            recordsToCreate[table] = []
+          }
+          recordsToCreate[table].push((raw: $FlowFixMe<RawRecord>))
+
           break
+        default:
+          break
+      }
+    })
+
+    // We're doing a second pass, because batch insert is much faster in Loki
+    Object.entries(recordsToCreate).forEach((args: any) => {
+      const [table, raws]: [TableName<any>, RawRecord[]] = args
+      const shouldRebuildIndexAfterIndex = raws.length >= 1000 // only profitable for large inserts
+      this.loki.getCollection(table).insert(raws, shouldRebuildIndexAfterIndex)
+
+      raws.forEach(raw => {
+        this.markAsCached(table, raw.id)
+      })
+    })
+
+    operations.forEach(operation => {
+      const [type, table, rawOrId] = operation
+      switch (type) {
         case 'update':
-          this.update(table, raw)
+          this.update(table, (rawOrId: $FlowFixMe<RawRecord>))
           break
         case 'markAsDeleted':
-          this.markAsDeleted(table, raw.id)
+          this.markAsDeleted(table, (rawOrId: $FlowFixMe<RecordId>))
           break
         case 'destroyPermanently':
-          this.destroyPermanently(table, raw.id)
+          this.destroyPermanently(table, (rawOrId: $FlowFixMe<RecordId>))
           break
         default:
           break
@@ -327,8 +348,8 @@ export default class LokiExecutor {
     logger.log(`[DB][Worker] Migration successful`)
   }
 
-  _executeCreateTableMigration({ name, columns }: CreateTableMigrationStep): void {
-    this._addCollection({ name, columns })
+  _executeCreateTableMigration({ schema }: CreateTableMigrationStep): void {
+    this._addCollection(schema)
   }
 
   _executeAddColumnsMigration({ table, columns }: AddColumnsMigrationStep): void {