+++ src/adapters/type.js
@@ -1,15 +1,19 @@
 // @flow
 
-import type Query from '../Query'
+import type { SerializedQuery } from '../Query'
 import type { TableName, AppSchema } from '../Schema'
 import type { SchemaMigrations } from '../Schema/migrations'
-import type Model, { RecordId } from '../Model'
+import type { RecordId } from '../Model'
 import type { RawRecord } from '../RawRecord'
 
 export type CachedFindResult = RecordId | ?RawRecord
 export type CachedQueryResult = Array<RecordId | RawRecord>
 export type BatchOperationType = 'create' | 'update' | 'markAsDeleted' | 'destroyPermanently'
-export type BatchOperation = [BatchOperationType, Model]
+export type BatchOperation =
+  | ['create', TableName<any>, RawRecord]
+  | ['update', TableName<any>, RawRecord]
+  | ['markAsDeleted', TableName<any>, RecordId]
+  | ['destroyPermanently', TableName<any>, RecordId]
 
 export interface DatabaseAdapter {
   schema: AppSchema;
@@ -20,10 +24,10 @@ export interface DatabaseAdapter {
   find(table: TableName<any>, id: RecordId): Promise<CachedFindResult>;
 
   // Fetches matching records. Should not send raw object if already cached in JS
-  query<T: Model>(query: Query<T>): Promise<CachedQueryResult>;
+  query(query: SerializedQuery): Promise<CachedQueryResult>;
 
   // Counts matching records
-  count<T: Model>(query: Query<T>): Promise<number>;
+  count(query: SerializedQuery): Promise<number>;
 
   // Executes multiple prepared operations
   batch(operations: BatchOperation[]): Promise<void>;