+++ src/utils/common/randomId/index.js
@@ -1,20 +1,21 @@
 // @flow
 
-import { join, times } from 'rambdax'
-
 // Only numers and letters for human friendliness
 const alphabet = '0123456789abcdefghijklmnopqrstuvwxyz'
+const alphabetLength = alphabet.length
 const idLength = 16
-
-const randomCharacter = () => {
-  const random = Math.floor(Math.random() * alphabet.length)
-  return alphabet[random]
-}
 
 // Note: for explanation of generating record IDs on the client side, see:
 // https://github.com/Nozbe/WatermelonDB/issues/5#issuecomment-442046292
 const randomId = (): string => {
-  return join('', times(randomCharacter, idLength))
+  let id = ''
+  for (let i = 0; i < idLength / 2; i += 1) {
+    const random = Math.floor(Math.random() * alphabetLength * alphabetLength)
+    id += alphabet[Math.floor(random / alphabetLength)]
+    id += alphabet[random % alphabetLength]
+  }
+
+  return id
 }
 
 let generator = () => randomId()